+++ server/server.go
@@ -58,6 +58,7 @@ func SetDefaults() {
 	viper.SetDefault("email.from", "support@example.com")
 	viper.SetDefault("server.secure_cookies", true)
 	viper.SetDefault("server.session_idle_timeout", 900)
+	viper.SetDefault("server.auth_debug", false)
 	viper.SetDefault("server.listen", "0.0.0.0:8866")
 	viper.SetDefault("server.read_timeout", 30)
 	viper.SetDefault("server.write_timeout", 30)